% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_pollen}
\alias{aggregate_pollen}
\title{Aggregate Hourly Data into Daily}
\usage{
aggregate_pollen(data, level = "daily")
}
\arguments{
\item{data}{Data Frame containing hourly concentrations}
}
\description{
Aggregate Hourly Data into Daily
}
