% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_kable}
\alias{create_kable}
\title{Create a Comparison Plot of Pollen Concentrations}
\usage{
create_kable(data, title = "")
}
\arguments{
\item{data}{A data frame that contains pollen data}

\item{title}{The title for the table}
}
\value{
a kable object
}
\description{
Create a Comparison Plot of Pollen Concentrations
}
