% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_data_cosmo}
\alias{import_data_cosmo}
\title{Retrieve Pollendata from txt-file (COSMO output with Fieldextra)}
\usage{
import_data_cosmo(datapath, type)
}
\arguments{
\item{datapath}{File path to the txt-file containing the data}

\item{type}{Hirst, Cosmo or Assim}
}
\description{
Retrieve Pollendata from txt-file (COSMO output with Fieldextra)
}
