% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_data_dwh}
\alias{import_data_dwh}
\title{Retrieve Pollendata from txt-file (DWH)}
\usage{
import_data_dwh(datapath)
}
\arguments{
\item{datapath}{File path to the txt-file containing the data}
}
\description{
Retrieve Pollendata from txt-file (DWH)
}
