% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_pollen}
\alias{impute_pollen}
\title{Impute Hourly Data}
\usage{
impute_pollen(
  data,
  min_date = min(data$datetime),
  max_date = max(data$datetime),
  taxon = unique(data$taxon)
)
}
\arguments{
\item{data}{Data Frame containing hourly concentrations}

\item{min_date}{Beginning of the timeseries in datetime format}

\item{max_date}{End of the timeseries in datetime format}

\item{taxon}{Unique taxon to impute each one seperately}
}
\description{
Impute Hourly Data
}
