% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_comb}
\alias{plot_comb}
\title{Create a Comparison Plot of Pollen Concentrations}
\usage{
plot_comb(data_plot, taxon, station, resolution, rm_zeros, combined, plot_dwh)
}
\arguments{
\item{data_plot}{A dataframe containing pollen concentrations}

\item{taxon}{The selected Pollen Type}

\item{station}{The selected Pollen Type}

\item{resolution}{Temporal resolution c("daily" or "hourly")}

\item{rm_zeros}{Should zero/NA pollen measurements be removed from the plot}

\item{combined}{Return the combined plots with title or seperate plots}

\item{plot_dwh}{Should the Measurements from the DWH be plotted}
}
\value{
A list of ggplots or a combined ggplot
}
\description{
Create a Comparison Plot of Pollen Concentrations
}
