% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_hours}
\alias{shift_hours}
\title{Shift the hours 15h backwards, as cosmo updates the phenology every day at 14h}
\usage{
shift_hours(data)
}
\arguments{
\item{data}{Data Frame containing hourly values}
}
\description{
Shift the hours 15h backwards, as cosmo updates the phenology every day at 14h
}
